Refactor the widget implementation to improve readability and maintainability.
The main changes are:
- Moved each widget layout to its own file in the `layouts` directory.
- Moved each widget component to its own file in the `components` directory.
- Removed unused components and code.
- Renamed components for clarity.
- The main `PixelPlayGlanceWidget.kt` file is now much smaller and cleaner.
