+++ app/src/main/java/com/theveloper/pixelplay/ui/glancewidget/layouts/VeryThinWidgetLayout.kt
@@ -1,0 +1,75 @@
+package com.theveloper.pixelplay.ui.glancewidget.layouts
+
+import android.content.Context
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import androidx.glance.GlanceModifier
+import androidx.glance.GlanceTheme
+import androidx.glance.appwidget.cornerRadius
+import androidx.glance.background
+import androidx.glance.layout.Alignment
+import androidx.glance.layout.Box
+import androidx.glance.layout.Row
+import androidx.glance.layout.Spacer
+import androidx.glance.layout.fillMaxHeight
+import androidx.glance.layout.fillMaxSize
+import androidx.glance.layout.padding
+import androidx.glance.layout.width
+import androidx.glance.unit.ColorProvider
+import com.theveloper.pixelplay.ui.glancewidget.components.AlbumArtImageGlance
+import com.theveloper.pixelplay.ui.glancewidget.components.NextButtonGlance
+import com.theveloper.pixelplay.ui.glancewidget.components.PlayPauseButtonGlance
+
+@Composable
+fun VeryThinWidgetLayout(
+    modifier: GlanceModifier,
+    title: String,
+    backgroundColor: ColorProvider,
+    bgCornerRadius: Dp,
+    artist: String,
+    albumArtBitmapData: ByteArray?,
+    isPlaying: Boolean,
+    textColor: ColorProvider,
+    context: Context
+) {
+    val secondaryColor = GlanceTheme.colors.secondaryContainer
+    val onSecondaryColor = GlanceTheme.colors.onSecondaryContainer
+    val primaryContainerColor = GlanceTheme.colors.primaryContainer
+    val onPrimaryContainerColor = GlanceTheme.colors.onPrimaryContainer
+
+    Box(
+        modifier = modifier.background(backgroundColor).cornerRadius(bgCornerRadius)
+            .padding(12.dp)
+    ) {
+        Row(
+            modifier = GlanceModifier.fillMaxSize().cornerRadius(bgCornerRadius),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalAlignment = Alignment.Horizontal.CenterHorizontally
+        ) {
+            AlbumArtImageGlance(
+                modifier = GlanceModifier.defaultWeight().fillMaxHeight(),
+                bitmapData = albumArtBitmapData,
+                context = context,
+                cornerRadius = bgCornerRadius
+            )
+            Spacer(GlanceModifier.width(8.dp))
+            PlayPauseButtonGlance(
+                modifier = GlanceModifier.defaultWeight().fillMaxHeight(),
+                backgroundColor = primaryContainerColor,
+                iconColor = onPrimaryContainerColor,
+                isPlaying = isPlaying,
+                iconSize = 26.dp,
+                cornerRadius = 10.dp
+            )
+            Spacer(GlanceModifier.width(8.dp))
+            NextButtonGlance(
+                modifier = GlanceModifier.defaultWeight().fillMaxHeight(),
+                iconColor = onSecondaryColor,
+                iconSize = 26.dp,
+                backgroundColor = secondaryColor,
+                cornerRadius = 10.dp
+            )
+        }
+    }
+}
