+++ app/src/main/java/com/theveloper/pixelplay/ui/glancewidget/utils/DurationFormatter.kt
@@ -1,0 +1,9 @@
+package com.theveloper.pixelplay.ui.glancewidget.utils
+
+internal fun formatDurationGlance(millis: Long): String {
+    if (millis < 0) return "00:00"
+    val totalSeconds = millis / 1000
+    val minutes = totalSeconds / 60
+    val seconds = totalSeconds % 60
+    return String.format("%02d:%02d", minutes, seconds)
+}
